requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
matplotlib==3.7.2
joblib==1.3.2
faker==19.6.2
python-dotenv==1.0.0
